requirements.txt
langchain
openai
faiss-cpu
pypdf
tiktoken
fastapi
uvicorn
python-dotenv